@@ -1,6 +1,6 @@
 import org.domnikl.gaia.FritzBox
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class FritzBoxTest {
     @Test
