@@ -1,7 +1,7 @@
 import kotlinx.coroutines.runBlocking
 import org.domnikl.gaia.TriggeringQueue
-import org.junit.Assert.assertEquals
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 class TriggeringQueueTest {
     private var l = emptyList<Float>()
